Port fixes from private repo

+++ vigpol/policer_config.c
@@ -16,10 +16,10 @@
 #include "nf-util.h"
 #include "nf-log.h"
 
-const uint32_t DEFAULT_LAN = 1;
-const uint32_t DEFAULT_WAN = 0;
-const uint32_t DEFAULT_RATE = 1000000; // 1MB/s
-const uint32_t DEFAULT_BURST = 100000; // 100kB
+const uint16_t DEFAULT_LAN = 1;
+const uint16_t DEFAULT_WAN = 0;
+const uint64_t DEFAULT_RATE = 1000000; // 1MB/s
+const uint64_t DEFAULT_BURST = 100000; // 100kB
 const uint32_t DEFAULT_CAPACITY = 128; // IPs
 
 #define PARSE_ERROR(format, ...)                                               \
@@ -99,9 +99,9 @@ void nf_config_usage(void) {
           "\t--wan <device>: WAN device,"
           " default: %" PRIu16 ".\n"
           "\t--rate <rate>: policer rate in bytes/s,"
-          " default: %" PRIu32 ".\n"
+          " default: %" PRIu64 ".\n"
           "\t--burst <size>: policer burst size in bytes,"
-          " default: %" PRIu32 ".\n"
+          " default: %" PRIu64 ".\n"
           "\t--capacity <n>: policer table capacity,"
           " default: %" PRIu32 ".\n",
           DEFAULT_LAN, DEFAULT_WAN, DEFAULT_RATE, DEFAULT_BURST,
@@ -111,10 +111,10 @@ void nf_config_usage(void) {
 void nf_config_print(void) {
   NF_INFO("\n--- Policer Config ---\n");
 
-  NF_INFO("LAN Device: %" PRIu32, config.lan_device);
-  NF_INFO("WAN Device: %" PRIu32, config.wan_device);
-  NF_INFO("Rate: %" PRIu32, config.rate);
-  NF_INFO("Burst: %" PRIu32, config.burst);
+  NF_INFO("LAN Device: %" PRIu16, config.lan_device);
+  NF_INFO("WAN Device: %" PRIu16, config.wan_device);
+  NF_INFO("Rate: %" PRIu64, config.rate);
+  NF_INFO("Burst: %" PRIu64, config.burst);
   NF_INFO("Capacity: %" PRIu16, config.dyn_capacity);
 
   NF_INFO("\n--- ------ ------ ---\n");
